Fix issues with scaling the container

+++ scripts/start.sh
@@ -47,7 +47,7 @@ EOF
     # Change  ownership to mapproxy user and mapproxy group
     ###
 
-    mkdir -p "${MAPPROXY_DATA_DIR}" /settings "${MULTI_MAPPROXY_DATA_DIR}" /root/.aws
+    mkdir -p "${MAPPROXY_DATA_DIR}" /settings "${MULTI_MAPPROXY_DATA_DIR}" /root/.aws ${MAPPROXY_APP_DIR}
     if [[ "${RECREATE_DATADIR}" =~ [Tt][Rr][Uu][Ee] ]]; then
         rm -rf "${MULTI_MAPPROXY_DATA_DIR}"/* "${MAPPROXY_DATA_DIR}"/*
     fi
@@ -61,21 +61,21 @@ EOF
     fi
     # Generate S3 configurations
     if [[ ${ENABLE_S3_CACHE} =~ [Tt][Rr][Uu][Ee] ]];then
-      export AWS_CONFIG_FILE AWS_SHARED_CREDENTIALS_FILE AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY S3_DEFAULT_PROFILE S3_DEFAULT_REGION S3_DEFAULT_OUTPUT
       configure_s3_cache
     fi
-    chown -R mapproxy:mapproxy "${MAPPROXY_DATA_DIR}" "${MULTI_MAPPROXY_DATA_DIR}" /settings /scripts/ root/.aws
+    chown -R mapproxy:mapproxy "${MAPPROXY_DATA_DIR}" "${MULTI_MAPPROXY_DATA_DIR}" /settings /scripts/ root/.aws ${MAPPROXY_APP_DIR}
 
     # Check if uwsgi configuration exists
     function uwisgi_config (){
        DATA_PATH=$1
        if [[ ! -f /settings/uwsgi.ini ]]; then
-          echo -e "\e[32m No custom uwsgi.ini file, will setup using default one provided at https://github.com/kartoza/docker-mapproxy/blob/master/uwsgi.ini \033[0m"
+          echo -e "\e[32m No custom uwsgi.ini file, setup using default one from  \033[0m \e[1;31m https://github.com/kartoza/docker-mapproxy/blob/master/uwsgi.ini \033[0m"
           # If it doesn't exists, copy from /mapproxy directory if exists
           if [[ -f ${DATA_PATH}/uwsgi.ini ]]; then
             cp -f "${DATA_PATH}"/uwsgi.ini /settings/uwsgi.ini
           else
             # default value
+            export CONFIG_DATA_PATH PROCESSES CHEAPER THREADS MAPPROXY_USER_ID MAPPROXY_GROUP_ID MAPPROXY_APP_DIR
             envsubst < /settings/uwsgi.default.ini > /settings/uwsgi.ini
           fi
         fi
@@ -86,11 +86,11 @@ EOF
     base_config_generator "${CONFIG_DATA_PATH}"
     pushd "${CONFIG_DATA_PATH}" || exit
 
-    if [[ ! -f "${CONFIG_DATA_PATH}"/app.py ]];then
-        mapproxy-util create -t wsgi-app -f "${CONFIG_DATA_PATH}"/mapproxy.yaml "${CONFIG_DATA_PATH}"/app.py
+    if [[ ! -f "${MAPPROXY_APP_DIR}"/app.py ]];then
+        mapproxy-util create -t wsgi-app -f "${CONFIG_DATA_PATH}"/mapproxy.yaml "${MAPPROXY_APP_DIR}"/app.py
     else
-        rm "${CONFIG_DATA_PATH}"/app.py
-        mapproxy-util create -t wsgi-app -f "${CONFIG_DATA_PATH}"/mapproxy.yaml "${CONFIG_DATA_PATH}"/app.py
+        rm "${MAPPROXY_APP_DIR}"/app.py
+        mapproxy-util create -t wsgi-app -f "${CONFIG_DATA_PATH}"/mapproxy.yaml "${MAPPROXY_APP_DIR}"/app.py
     fi
     if [[ -f "${CONFIG_DATA_PATH}"/full_example.yaml ]];then
         rm "${CONFIG_DATA_PATH}"/full_example.yaml 2> /dev/null || true
@@ -104,20 +104,27 @@ EOF
         if [[ -f /settings/log.ini ]];then
           cp /settings/log.ini "${CONFIG_DATA_PATH}"/log.ini
         else
+            # Always create a new log.ini
             if [[ ! -f "${CONFIG_DATA_PATH}"/log.ini ]];then
-                mapproxy-util create -t log-ini "${CONFIG_DATA_PATH}"/log.ini
+                mapproxy-util create -t log-ini --force "${CONFIG_DATA_PATH}"/log.ini
+            else
+                rm "${CONFIG_DATA_PATH}"/log.ini
+                mapproxy-util create -t log-ini --force "${CONFIG_DATA_PATH}"/log.ini
             fi
         fi
         #TODO - Fix sed to replace all commands in one go
-        sed -i 's/%(here)s/${CONFIG_DATA_PATH}/g' "${CONFIG_DATA_PATH}"/log.ini
-        envsubst < "${CONFIG_DATA_PATH}"/log.ini > "${CONFIG_DATA_PATH}"/log.ini.bak
-        mv "${CONFIG_DATA_PATH}"/log.ini.bak "${CONFIG_DATA_PATH}"/log.ini
+        # Add custom logic if it doesn't come from a user defined one
+        if [[ ! -f /settings/log.ini ]];then
+          sed -i 's/%(here)s/${CONFIG_DATA_PATH}/g' "${CONFIG_DATA_PATH}"/log.ini
+          envsubst < "${CONFIG_DATA_PATH}"/log.ini > "${CONFIG_DATA_PATH}"/log.ini.bak
+          mv "${CONFIG_DATA_PATH}"/log.ini.bak "${CONFIG_DATA_PATH}"/log.ini
+        fi
     fi
 
     # Add logic to reload the app file
     RELOAD_LOCKFILE=/settings/.app.lock
     if [[ ! -f ${RELOAD_LOCKFILE} ]];then
-      sed -i 's/\(, reloader=True\)*'\)'/, reloader=True\)/g' "${CONFIG_DATA_PATH}"/app.py
+      sed -i 's/\(, reloader=True\)*'\)'/, reloader=True\)/g' "${MAPPROXY_APP_DIR}"/app.py
       touch ${RELOAD_LOCKFILE}
     fi
     # Enable logging module
@@ -128,8 +135,8 @@ from logging.config import fileConfig
 import os.path
 fileConfig(r'${CONFIG_DATA_PATH}/log.ini', {'here': os.path.dirname(__file__)})
       " > /tmp/log.py
-      cat "${CONFIG_DATA_PATH}"/app.py >> /tmp/log.py
-      mv  /tmp/log.py  "${CONFIG_DATA_PATH}"/app.py
+      cat "${MAPPROXY_APP_DIR}"/app.py >> /tmp/log.py
+      mv  /tmp/log.py  "${MAPPROXY_APP_DIR}"/app.py
     fi
     }
 
@@ -141,7 +148,8 @@ fileConfig(r'${CONFIG_DATA_PATH}/log.ini', {'here': os.path.dirname(__file__)})
         exec gosu mapproxy uwsgi --ini /settings/uwsgi.ini
     elif [[ ${PRODUCTION} =~ [Tt][Rr][Uu][Ee] ]] && [[ ${MULTI_MAPPROXY} =~ [Tt][Rr][Uu][Ee] ]]; then
         uwisgi_config "${CONFIG_DATA_PATH}"
-        envsubst < /multi_mapproxy.py > "${CONFIG_DATA_PATH}"/app.py
+        export MULTI_MAPPROXY_DATA_DIR ALLOW_LISTING
+        envsubst < /multi_mapproxy.py > "${MAPPROXY_APP_DIR}"/app.py
         make_logs
         exec gosu mapproxy uwsgi --ini /settings/uwsgi.ini
     else
